Extract catalog_privacy.py + routing_candidates.py, completing #38 split
Final step of decomposing the catalog.py god-module. This lifts out the two
remaining cohesive concerns:

- catalog_privacy.py (~70 lines): the 6 privacy-posture tier functions
  (provider/model/endpoint privacy tier + zero-data-retention/policy/policy_url).
  Pure leaf — imports only catalog_data.
- routing_candidates.py (~415 lines): the meta-model candidate-selection surface
  (auto/free/cheap/fast/monitor/eu/zdr/e2e/socrates/meta candidate_models,
  the meta router, InvalidAutoModelOrder, and their private helpers). Imports the
  registry (MODELS/MODEL_ENDPOINTS), catalog_data constants, and
  catalog_privacy.endpoint_privacy_tier.

catalog.py drops from ~1000 to 620 lines and re-exports both modules' full
public surface, so every `from trusted_router.catalog import ...` caller keeps
working unchanged.

Both blocks were contiguous and self-contained: the moved function bodies are
byte-identical, the routing region references ZERO catalog-STAY functions (no
cycle — verified: privacy + routing import without loading catalog), and the
public surface is unchanged (dir(catalog) non-underscore names = 124, same as
before). A behavioral-equivalence snapshot over every candidate function, meta
model, provider/endpoint privacy tier, model_open_weights, meta route kind, and
validate_auto_model_order is byte-for-byte identical pre/post split.

Full suite: 1193 passed, 33 skipped. ruff + mypy clean.
